+++ CHANGELOG.md
@@ -1,17 +1,26 @@
-## 0.3.0
+## 0.3.2+1
+
+- Added Number Ticker Widget for displaying real-time changes of a number.
+- Introduced extension for percentage calculation.
+
+## 0.3.1
+
 - Added combineWordsWithSpecialChar Extension for String
 - Bug fixed in svgIconImage()
 
 ## 0.2.0
+
 - Added GroupBy Extension for List
 - Added wrapWithBrackets Extension for String
 - Added New Enmus
 
 ## 0.1.0
+
 - Added View All Widget
 - Added Circle Widget
 
 ## 0.0.5
+
 - Added Name Initial Widget
 - Added Custom Divider (4 Different types of divider)
 - Added FadeIn Animation (Can be used with any widgets)